+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/util/UserPresence.kt
@@ -7,6 +7,7 @@ import net.folivo.trixnity.client.user
 import net.folivo.trixnity.core.model.RoomId
 import net.folivo.trixnity.core.model.events.m.Presence
 import net.folivo.trixnity.core.model.events.m.PresenceEventContent
+import net.folivo.trixnity.core.model.events.m.room.Membership
 
 interface UserPresence {
     fun presentEventContentFlow(
@@ -26,7 +27,9 @@ class UserPresenceImpl(
     ): Flow<PresenceEventContent?> {
         return combine(
             matrixClient.user.userPresence,
-            directRoom.getUsers(matrixClient, roomId)
+            directRoom.getUsersWithMembership(matrixClient, roomId).map { users ->
+                users.filter { (_, membership) -> membership == Membership.JOIN }.map { (id, _) -> id }
+            }
         ) { userPresence, otherUsers ->
             otherUsers.firstOrNull()?.let { userId ->
                 userPresence[userId]?.let { presence ->